refactor(settings): move agent hotkey to hotkeys section
